Open Source Mathematica ® Code Multiconductor Modeling and Analysis Microstrip Circuits
The Mathematica Code (Ardalan)  is based on Riddle, Ardalan , Suh, 1988  and has no approximations. Although the code provided is for a 2 Conductor Coupled Microstrip circuit it is written for an arbitrary number of conductors. The key is to generate the Per Unit Length Parameters using the Mathematica Tool and pasting them in ( for Matrices LL and Cap). Also set the number of conductors ( NN). Note that you must decide which result you want to examine. Look at the Vx vector which is the calculated voltage at each node at a distance x from the source To calculate S14 ( Far End Crosstalk) you need Vx for x=0 and Vx fo x=L. Two step process. Unless you set the source impedance to zero for the generator.